# coreApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Johnny Cms Api",
        "description": "Johnny Cms Api ",
        "contact": {
            "email": "johnny_login@163.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/newArticeClass": {
            "post": {
                "tags": [
                    "Cms 文章/新闻分类"
                ],
                "summary": "newArticeClass 创建文章分类",
                "description": "newArticeClass 创建文章分类",
                "operationId": "Module",
                "parameters": [
                    {
                        "name": "sid",
                        "in": "path",
                        "description": "所属id",
                        "required": true,
                        "schema": {
                            "type": "Interger"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "分类名称",
                        "required": true,
                        "schema": {
                            "type": "String"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/setArticeClass": {
            "post": {
                "tags": [
                    "Cms 文章/新闻分类"
                ],
                "summary": "setArticeClass 修改文章分类",
                "description": "setArticeClass 修改文章分类",
                "operationId": "Module",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "文章id",
                        "required": true,
                        "schema": {
                            "type": "Interger"
                        }
                    },
                    {
                        "name": "sid",
                        "in": "path",
                        "description": "所属id",
                        "required": true,
                        "schema": {
                            "type": "Interger"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "分类名称",
                        "required": true,
                        "schema": {
                            "type": "String"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/delArticeClass": {
            "post": {
                "tags": [
                    "Cms 文章/新闻分类"
                ],
                "summary": "delArticeClass 删除文章分类",
                "description": "delArticeClass 删除文章分类",
                "operationId": "Module",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "文章id",
                        "required": true,
                        "schema": {
                            "type": "Interger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/newProductClass": {
            "post": {
                "tags": [
                    "Cms 商品分类"
                ],
                "summary": "创建商品分类",
                "description": "创建商品分类",
                "operationId": "Module",
                "parameters": [
                    {
                        "name": "sid",
                        "in": "path",
                        "description": "所属ID",
                        "required": true,
                        "schema": {
                            "type": "Integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "分类名称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/setProductClass": {
            "post": {
                "tags": [
                    "Cms 商品分类"
                ],
                "summary": "更新分类名称",
                "description": "更新分类名称",
                "operationId": "Module",
                "parameters": [
                    {
                        "name": "sid",
                        "in": "path",
                        "description": "所属ID",
                        "required": true,
                        "schema": {
                            "type": "Integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "分类名称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "目标ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/delProductClass": {
            "post": {
                "tags": [
                    "Cms 商品分类"
                ],
                "summary": "删除分类",
                "description": "删除分类",
                "operationId": "Module",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "目标ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    }
}